+++ src/extensions/building/buildingext_hooks.cpp
@@ -2362,6 +2362,84 @@ past_flame_spawn:
 }
 
 
+BuildingClass* Find_Best_Alternative_Factory(BuildingClass* this_ptr, FootClass* exiting_object)
+{
+    int closest_distance = INT_MAX;
+    BuildingClass* closest_match = nullptr;
+
+    for (int i = 0; i < Buildings.Count(); i++) {
+        BuildingClass* bldg = Buildings[i];
+
+        if (bldg->House == this_ptr->House && bldg != this_ptr && bldg->Mission == MISSION_GUARD && !bldg->Factory && bldg->Class->ToBuild == this_ptr->Class->ToBuild && bldg->Is_Powered_On()) {
+            // Original TS code, left here for reference. Was part of the above condition
+            // if (bldg->Class != this_ptr->Class)
+            //     continue;
+
+            const TechnoTypeClass* technotype = exiting_object->Techno_Type_Class();
+
+            // Check ownable, so only factories of a faction that owns the object can
+            // build the object
+            if ((bldg->Class->Get_Ownable() & technotype->Get_Ownable()) == 0) {
+                continue;
+            }
+
+            // Check ownable, so only factories of a faction that owns the object can
+            // build the object
+            if ((bldg->Class->Get_Ownable() & exiting_object->Techno_Type_Class()->Get_Ownable()) == 0) {
+                continue;
+            }
+
+            // All checks have passed. Check the distance to find the closest factory to exit from.
+            int distance = this_ptr->Distance_To(bldg);
+            if (distance < closest_distance) {
+                closest_distance = distance;
+                closest_match = bldg;
+            }
+        }
+    }
+
+    return closest_match;
+}
+
+
+/**
+ *  Replaces the loop starting from 0x0042CAB9 to improve the alternative
+ *  war factory selection logic when a war factory is busy in 3 ways:
+ * 
+ *  1) The object can now exit from factory buildings of a different type
+ *     than what the object was originally produced from.
+ *  
+ *  2) The above takes speed type into account when finding building to exit from,
+ *     preventing ships from exiting from land-based factories and vice-versa.
+ * 
+ *  3) The logic prefers finding the closest rather than the "first" alternative
+ *  factory building.
+ * 
+ *  @author: Rampastring
+ */
+DEFINE_HOOK(0x0042CAB9, _BuildingClass_Exit_Object_Factory_Busy_Customized_Alternate_Factory_Seeking_Logic, 0)
+{
+    GET(BuildingClass*, this_ptr, ESI);
+    GET(FootClass*, exiting_object, EDI);
+
+    BuildingClass* best_alternative_building = Find_Best_Alternative_Factory(this_ptr, exiting_object);
+    if (best_alternative_building != nullptr) {
+
+        /**
+         *  Exit from the factory we found.
+         */
+        R->EBP(best_alternative_building);
+        return 0x0042CB28;
+    }
+
+    /**
+     *  We could not find an alternative factory to exit from,
+     *  exit the function and return 1.
+     */
+    return 0x0042CB16;
+}
+
+
 /**
  *  Main function for patching the hooks.
  */