+++ docs/Whats-New.md
@@ -76,6 +76,7 @@ New:
 - Add customizable wake animations (by CCHyper, ZivDero)
 - Replace DirectDraw with SDL (by ZivDero, tomsons26, CCHyper)
 - Make it possible to prevent buildings from repeatedly catching fire when rapidly switching between damage stages (by Rampastring)
+- Improve alternative factory selection when the primary factory is blocked (by Rampstring)
 
 
 Vinifera fixes: